Move to package-inferred system with more code separation by file
and one package defined per file.

Code is now split out to MAIN, GLOBALS, HTTP, CRYPTOGRAPHY, and TIME.

The test system is updated, with tests split out to HTTP, CRYPTOGRAPHY, and TIME... to be completed.

The refactoring made it clear that the PATH logic in CRYPTOGRAPHY:SIGNATURE ought to be moved to HTTP:POST-PRIVATE to share and simplify dependencies. Done in this commit.

File headers were simplified since the new package definitions are self-explanatory.

Version bumped from 0.0.1 to 0.0.2 in the ASD file.

Moving to one-package-per-file is arguably unnecessary given the current small size of this library, but I'm doing it to learn and prepare for larger projects in Common Lisp. I do like the increased discipline and domain separation of this approach.