Remove unnecessary QURI dependency in GLOBALS
and define constants for the API HTTP `scheme` and `host` which along with the paths can be used for building URIs using QURI:MAKE-URI for GET-PUBLIC and POST-PRIVATE in `src/http`.

Rename the "url" parameters to "path".

Use "+" instead of "*" earmuffs for constants whose values will not change between server restarts.